ci/tests: Refactor lines for waiting for instance state
